# contracts/openapi.json
{
  "components": {
    "schemas": {
      "AccountView": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "role"
        ],
        "title": "AccountView",
        "type": "object"
      },
      "AdminRevisionView": {
        "properties": {
          "creator": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Creator"
          },
          "decision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "media_type": {
            "title": "Media Type",
            "type": "string"
          },
          "metadata_version": {
            "title": "Metadata Version",
            "type": "integer"
          },
          "origin_url": {
            "title": "Origin Url",
            "type": "string"
          },
          "original_available": {
            "title": "Original Available",
            "type": "boolean"
          },
          "people": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "People"
          },
          "period": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "revision_id": {
            "format": "uuid",
            "title": "Revision Id",
            "type": "string"
          },
          "rights_usage_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rights Usage Note"
          },
          "segment_review_version": {
            "default": 0,
            "title": "Segment Review Version",
            "type": "integer"
          },
          "segments": {
            "items": {
              "$ref": "#/components/schemas/SegmentView"
            },
            "title": "Segments",
            "type": "array"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          },
          "source_id": {
            "format": "uuid",
            "title": "Source Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/TagView"
            },
            "title": "Tags",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "revision_id",
          "title",
          "creator",
          "origin_url",
          "rights_usage_note",
          "media_type",
          "tags",
          "segments",
          "original_available",
          "source_id",
          "status",
          "error_code",
          "decision",
          "sha256",
          "metadata_version"
        ],
        "title": "AdminRevisionView",
        "type": "object"
      },
      "AdminSourceView": {
        "properties": {
          "decision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision"
          },
          "origin_url": {
            "title": "Origin Url",
            "type": "string"
          },
          "revision_id": {
            "format": "uuid",
            "title": "Revision Id",
            "type": "string"
          },
          "source_id": {
            "format": "uuid",
            "title": "Source Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "source_id",
          "revision_id",
          "title",
          "origin_url",
          "status",
          "decision"
        ],
        "title": "AdminSourceView",
        "type": "object"
      },
      "AuthView": {
        "properties": {
          "csrf_token": {
            "title": "Csrf Token",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/AccountView"
          }
        },
        "required": [
          "user",
          "csrf_token"
        ],
        "title": "AuthView",
        "type": "object"
      },
      "Body_upload_source_api_v1_admin_sources_post": {
        "properties": {
          "creator": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Creator"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "origin_url": {
            "maxLength": 2000,
            "minLength": 8,
            "title": "Origin Url",
            "type": "string"
          },
          "rights_note": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rights Note"
          },
          "source_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "title": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "file",
          "origin_url",
          "title"
        ],
        "title": "Body_upload_source_api_v1_admin_sources_post",
        "type": "object"
      },
...
      "ChatDetailView": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "turns": {
            "items": {
              "$ref": "#/components/schemas/ChatTurnView"
            },
            "title": "Turns",
            "type": "array"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "updated_at",
          "turns"
        ],
        "title": "ChatDetailView",
        "type": "object"
      },
...
      "CreateAccountInput": {
        "properties": {
          "password": {
            "maxLength": 128,
            "minLength": 12,
            "title": "Password",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/Role",
            "default": "user"
          },
          "username": {
            "maxLength": 50,
            "minLength": 3,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "CreateAccountInput",
        "type": "object"
      },
...
      "DecisionInput": {
        "properties": {
          "evidence_url": {
            "maxLength": 2000,
            "minLength": 8,
            "title": "Evidence Url",
            "type": "string"
          },
          "original_file": {
            "default": false,
            "title": "Original File",
            "type": "boolean"
          },
          "provider_transfer": {
            "default": false,
            "title": "Provider Transfer",
            "type": "boolean"
          },
          "reason": {
            "maxLength": 2000,
            "minLength": 10,
            "title": "Reason",
            "type": "string"
          },
          "sensitivity_cleared": {
            "default": false,
            "title": "Sensitivity Cleared",
            "type": "boolean"
          },
          "user_text": {
            "default": false,
            "title": "User Text",
            "type": "boolean"
          }
        },
        "required": [
          "reason",
          "evidence_url"
        ],
        "title": "DecisionInput",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IntakeView": {
        "properties": {
          "revision_id": {
            "format": "uuid",
            "title": "Revision Id",
            "type": "string"
          },
          "source_id": {
            "format": "uuid",
            "title": "Source Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "source_id",
          "revision_id",
          "status"
        ],
        "title": "IntakeView",
        "type": "object"
      },
...
      "LoginInput": {
        "properties": {
          "password": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          },
          "username": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "LoginInput",
        "type": "object"
      },
...
      "RevokeInput": {
        "properties": {
          "reason": {
            "maxLength": 2000,
            "minLength": 10,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "RevokeInput",
        "type": "object"
      },
      "Role": {
        "enum": [
          "user",
          "admin"
        ],
        "title": "Role",
        "type": "string"
      },
...
      "SendInput": {
        "properties": {
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "starter_id": {
            "anyOf": [
              {
                "enum": [
                  "UC-01",
                  "UC-02",
                  "UC-03",
                  "UC-04",
                  "UC-05",
                  "UC-06"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Starter Id"
          },
          "text": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "text"
        ],
        "title": "SendInput",
        "type": "object"
      },
...
      "TagView": {
        "properties": {
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "value"
        ],
        "title": "TagView",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "CultureChamp API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/admin/accounts": {
      "post": {
        "operationId": "create_account_api_v1_admin_accounts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAccountInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Account",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/v1/admin/revisions/{revision_id}": {
      "get": {
        "operationId": "admin_revision_api_v1_admin_revisions__revision_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Revision Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminRevisionView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Revision",
        "tags": [
          "admin-sources"
        ]
      }
    },
    "/api/v1/admin/revisions/{revision_id}/approve": {
      "post": {
        "operationId": "approve_revision_api_v1_admin_revisions__revision_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Revision Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminRevisionView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve Revision",
        "tags": [
          "admin-sources"
        ]
      }
    },
...
    "/api/v1/admin/revisions/{revision_id}/revoke": {
      "post": {
        "operationId": "revoke_revision_api_v1_admin_revisions__revision_id__revoke_post",
        "parameters": [
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Revision Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminRevisionView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke Revision",
        "tags": [
          "admin-sources"
        ]
      }
    },
...
    "/api/v1/admin/sources": {
      "get": {
        "operationId": "admin_sources_api_v1_admin_sources_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "candidate",
                    "processing",
                    "review_pending",
                    "failed"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "decision",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "approve",
                    "revoke",
                    "none"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Decision"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "media_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "application/pdf",
                    "text/plain",
                    "text/csv",
                    "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Media Type"
            }
          },
          {
            "in": "query",
            "name": "tag_kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "region",
                    "people",
                    "period",
                    "topic",
                    "sensitivity"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag Kind"
            }
          },
          {
            "in": "query",
            "name": "tag_value",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag Value"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AdminSourceView"
                  },
                  "title": "Response Admin Sources Api V1 Admin Sources Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Sources",
        "tags": [
          "admin-sources"
        ]
      },
      "post": {
        "operationId": "upload_source_api_v1_admin_sources_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_source_api_v1_admin_sources_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntakeView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Source",
        "tags": [
          "admin-sources"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/chats/{chat_id}": {
      "delete": {
        "operationId": "delete_chat_api_v1_chats__chat_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "chat_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Chat Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Chat",
        "tags": [
          "chats"
        ]
      },
      "get": {
        "operationId": "chat_detail_api_v1_chats__chat_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "chat_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Chat Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatDetailView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chat Detail",
        "tags": [
          "chats"
        ]
      }
    },
    "/api/v1/chats/{chat_id}/messages": {
      "post": {
        "operationId": "send_message_api_v1_chats__chat_id__messages_post",
        "parameters": [
          {
            "in": "path",
            "name": "chat_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Chat Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatTurnView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Send Message",
        "tags": [
          "chats"
        ]
      }
    },
...
    "/api/v1/materials/{revision_id}": {
      "get": {
        "operationId": "material_detail_api_v1_materials__revision_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Revision Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Material Detail",
        "tags": [
          "materials"
        ]
      }
    },
...
  }
}